don't assume tarballs are sanely packaged
There's sadly a large number of packages in the AUR which fail at life
and can't manage to use a subdirectory matching the name of the package.
Rely on the fact that the AUR does some level of sanity checking on
uploads to make sure that one and only subdir exists and return this as
an outval from our archive extraction function.

People who do this need to go play in traffic.