Enforce LC_NUMERIC=C for talking to the AUR
yajl uses strtod for parsing doubles, making it locale-dependent. The
AUR always responds with "." as a decimal separator, breaking
conversions. The easy way to deal with this is to just fiddle with the
locale before talking to the AUR and then resetting to the
user-requested locale before we print results.